% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_plateinfo}
\alias{get_plateinfo}
\title{Return information about an MEA plate format}
\usage{
get_plateinfo(arrayname)
}
\arguments{
\item{arrayname}{Plate name}
}
\value{
A list storing information about that plate.
}
\description{
Return information about an MEA plate format
}
\details{
Given a plate name, return a list of information.
If the plate name is not recognised, an error is generated.
}
\author{
Stephen Eglen
}
