\name{has_network_spikes}
\alias{has_network_spikes}

\title{
A utility function to check if network spikes are detected.
}
\description{
For an returned object from calcualte.network.spikes, this function provides a utility checking if it contains any network spikes from any well.
}
\usage{
has_network_spikes(nspikes)
}
\arguments{
  \item{nspikes}{ The network spike object returned by calling calcualte.network.spikes.
}
}

\value{
Return a boolean value indicating if any network spikes are found from the network spikes object.
}


\keyword{network }
\keyword{spikes }
