\name{IGM.xyplot.network.spikes}
\alias{IGM.xyplot.network.spikes}
\title{
xyplot for network spikes at the plate level.
}
\description{
xyplot for network spikes at the plate level. It will detech the plate layout and plot individual wells according to plate layout. 
}
\usage{
IGM.xyplot.network.spikes(nspikes)
}

\arguments{
  \item{nspikes}{
  The returned obect from calling summary.network.spikes.
}
}

\value{
Return the handle of xyplot. 
}

\author{
Quanli Wang
}

\keyword{network spikes}
\keyword{ xyplot }
