\name{load_spikelist}
\alias{load_spikelist}

\title{
    Load Robject File
}

\description{
    Loads a previously saved Rdata file of a recording.
}
\usage{
load_spikelist(spk_data_file)
}

\arguments{
  \item{spk_data_file}{MEA recording Rdata file}
}
\value{loaded R object}

\keyword{Rdata}
\keyword{s}
