% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcs.R
\name{make_mcs100um_pos}
\alias{make_mcs100um_pos}
\title{Make the MCS 8x8 100um spacing .pos file}
\usage{
make_mcs100um_pos()
}
\description{
Make the MCS 8x8 100um spacing .pos file
}
\details{
This creates a file in the temp directory that can then be moved.
}
\author{
Stephen Eglen
}
