% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\docType{data}
\name{plateinfo}
\alias{plateinfo}
\title{MEA plate information}
\format{An object of class \code{list} of length 2.}
\usage{
plateinfo
}
\description{
MEA plate information
}
\keyword{datasets}
