% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sttc.R
\name{plot.sttcp}
\alias{plot.sttcp}
\title{Plot the STTCP}
\usage{
\method{plot}{sttcp}(x, ...)
}
\arguments{
\item{x}{A list containing the sttc}

\item{...}{Other arguments to pass to the plot function}
}
\value{
nothing.
}
\description{
Plot the STTCP
}
\details{
Plot the STTCP.
}
\author{
Stephen Eglen
}
