\name{plot_network_spikes}
\alias{plot_network_spikes}
\title{
Generic method for plotting network spikes.
}
\description{
The generic plotting function for network spikes.
}
\usage{
plot_network_spikes(ns, ...)
}
\arguments{
  \item{ns}{
    The network spike object returned after running calculate_network_spikes.
}
  \item{\dots}{
  Additional plotting options that is general to plot functions.
}
}

\value{
None. network spikes related plots will be generated in current plotting device.
}

\keyword{network spikes}
\keyword{plot}
