% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{publications}
\alias{publications}
\title{Show list of publications have used this package}
\usage{
publications()
}
\description{
Show list of publications have used this package
}
\details{
Simple wrapper function to show the publications.md file, documenting
papers that have used previous versions of this package.
In RStudio, the file will appear in its own window; otherwise it will
appear in a pager run within the R session
}
\author{
Stephen Eglen
}
