% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcs.R
\name{read_spikelist_mcs100um}
\alias{read_spikelist_mcs100um}
\title{Read in MCS 8x8 100um data spikes}
\usage{
read_spikelist_mcs100um(file, sample_rate = 25000)
}
\arguments{
\item{file}{File to read in}

\item{sample_rate}{Sample recording rate (default 25 Khz)}
}
\value{
S object
}
\description{
Read in MCS 8x8 100um data spikes
}
\details{
This uses the textreader routines for reading in MCS 8x8 arrays.
}
\author{
Stephen Eglen
}
