% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{run_TMcpp}
\alias{run_TMcpp}
\title{Compute STCC direct in Cpp}
\usage{
run_TMcpp(dt, start, end, spike_times_1, spike_times_2)
}
\arguments{
\item{dt}{bin width for}

\item{start}{start time in seconds}

\item{end}{end time in seconds}

\item{spike_times_1}{spike train 1}

\item{spike_times_2}{spike train 2}
}
\value{
STTC value
}
\description{
run the STTC code for a spike train (Cpp version)
}
\details{
Internal computation
}
\author{
Stephen Eglen
}
