% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{sttc_allspikes1}
\alias{sttc_allspikes1}
\title{Compute STTC for all pairs of spike trains}
\usage{
sttc_allspikes1(spikes, dt, beg, end)
}
\arguments{
\item{spikes}{List of spike trains}

\item{dt}{tiling window}

\item{beg}{start time}

\item{end}{end time}
}
\value{
Matrix of STTC values.  Upper diagonal matrix only; diagonal
elements should be 1.
}
\description{
Compute STTC for all unique pairs of spike trains
}
\details{
Return a matrix of all STTC values
}
\author{
Stephen Eglen
}
