% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{sttcp_ab}
\alias{sttcp_ab}
\title{Compute STTC profile for two spike trains}
\usage{
sttcp_ab(a, b, start, end, dt, tau_sep, tau_max)
}
\arguments{
\item{a}{Spike train 1}

\item{b}{Spike train 2}

\item{start}{Start time}

\item{end}{End time}

\item{dt}{coincidence window for STTC}

\item{tau_sep}{step size for tau in [-tau_max, +tau_max]}

\item{tau_max}{maximum tau value}
}
\value{
obj  An object of type "sttcp", containing the tau values and correlations.
}
\description{
Compute STTC profile for two spike trains
}
\details{
Compute the STTC profile for two spike trains using C++.
}
\author{
Stephen Eglen
}
