% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sttc.R
\name{plot.sttcp}
\alias{plot.sttcp}
\title{Simple plotting function to show STTC profile}
\usage{
\method{plot}{sttcp}(x, ...)
}
\arguments{
\item{x}{An sttcp object generated by the sttcp() function.}

\item{...}{Other plotting arguments to pass to plot()}
}
\value{
Nothing is returned.
}
\description{
Simple plotting function to show STTC profile
}
\details{
This is a simple plotting function to show the STTC profile.
Examples of its usage are given in the help page for sttcp.
}
\seealso{
sttcp
}
\author{
Stephen Eglen
}
