% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikelist_text.R
\name{read_spikelist_text}
\alias{read_spikelist_text}
\title{Construct an MEA object by importing from text files.}
\usage{
read_spikelist_text(spike_text_file, channel_text_file, chem_info, array)
}
\arguments{
\item{spike_text_file}{CSV with Spike time information}

\item{channel_text_file}{CSV with information regarding MEA layout}

\item{chem_info}{Optional chemical information (currently unused)}

\item{array}{Name of array}
}
\value{
S the MEA object
}
\description{
Construct an MEA object by importing from text files.
}
\details{
This function can be used to read in spike times from text files.
}
\examples{

}
\author{
Stephen Eglen
}
